Refactor imports to use centralized types file for improved code organization
