+++ src/app/flow/questions.ts
@@ -1,12 +1,4 @@
-export type Answer = string | string[]; // Updated to support multiple values
-
-export interface Question {
-  id: string;
-  text: string;
-  options: { label: string; value: string }[];
-  multiple?: boolean; // Indicates if multiple values can be selected
-  dependsOn?: (answers: Record<string, Answer>) => boolean;
-}
+import { Question } from "./types";
 
 export const questions: Question[] = [
   // Base Questions (Single Selection)
