+++ src/app/flow/recommendations.ts
@@ -1,49 +1,51 @@
+import { Answer } from "./types";
+
 export interface MacBookRecommendation {
   model: string;
   reason: string;
 }
 
 export const getRecommendation = (
-  answers: Record<string, string>
+  answers: Record<string, Answer>,
 ): MacBookRecommendation => {
   const { usage, portability, battery } = answers;
 
-  if (usage === 'video') {
+  if (usage === "video") {
     return {
-      model: 'MacBook Pro 16-inch (M4 Max)',
-      reason: 'Best for video editing with high performance and large screen.',
+      model: "MacBook Pro 16-inch (M4 Max)",
+      reason: "Best for video editing with high performance and large screen.",
     };
   }
 
-  if (usage === 'programming') {
-    if (portability === 'yes') {
+  if (usage === "programming") {
+    if (portability === "yes") {
       return {
-        model: 'MacBook Air 15-inch (M4)',
+        model: "MacBook Air 15-inch (M4)",
         reason:
-          'Great for portability while offering good performance for coding.',
+          "Great for portability while offering good performance for coding.",
       };
     }
     return {
-      model: 'MacBook Pro 14-inch (M4 Pro)',
-      reason: 'Balances power and portability for developers.',
+      model: "MacBook Pro 14-inch (M4 Pro)",
+      reason: "Balances power and portability for developers.",
     };
   }
 
-  if (usage === 'general') {
-    if (battery === 'yes') {
+  if (usage === "general") {
+    if (battery === "yes") {
       return {
-        model: 'MacBook Air 13-inch (M4)',
-        reason: 'Best for general use with long battery life.',
+        model: "MacBook Air 13-inch (M4)",
+        reason: "Best for general use with long battery life.",
       };
     }
     return {
-      model: 'MacBook Pro 14-inch (M4)',
-      reason: 'Slightly more powerful for general tasks.',
+      model: "MacBook Pro 14-inch (M4)",
+      reason: "Slightly more powerful for general tasks.",
     };
   }
 
   return {
-    model: 'MacBook Air 13-inch (M4)',
-    reason: 'A budget-friendly option for everyday tasks.',
+    model: "MacBook Air 13-inch (M4)",
+    reason: "A budget-friendly option for everyday tasks.",
   };
 };
